Refactor: move loop hoist analysis helper
Relocated the array-length hoist detection utilities out of src/plugin/src/printer/print.js and into src/plugin/src/printer/optimizations/loop-size-hoisting.js so the bulky AST analysis lives with other printer optimizations instead of inside the main print pipeline.
