+++ src/plugin/src/printer/optimizations/loop-size-hoisting.js
@@ -1,0 +1,154 @@
+// Helpers for detecting and naming cached loop size variables.
+// This logic analyses the AST rather than producing Prettier docs, so it lives
+// alongside other printer optimizations instead of the main print pipeline.
+
+const SIZE_RETRIEVAL_FUNCTION_SUFFIXES = new Map([
+    ["array_length", "len"],
+    ["ds_list_size", "size"],
+    ["ds_map_size", "size"],
+    ["ds_grid_width", "width"],
+    ["ds_grid_height", "height"],
+]);
+
+function getArrayLengthHoistInfo(node, sizeFunctionSuffixes = SIZE_RETRIEVAL_FUNCTION_SUFFIXES) {
+    if (!node || node.type !== "ForStatement") {
+        return null;
+    }
+
+    const test = node.test;
+    if (!test || test.type !== "BinaryExpression") {
+        return null;
+    }
+
+    if (!test.operator || !["<", "<="].includes(test.operator)) {
+        return null;
+    }
+
+    const callExpression = test.right;
+    if (!callExpression || callExpression.type !== "CallExpression") {
+        return null;
+    }
+
+    const callee = callExpression.object;
+    if (!callee || callee.type !== "Identifier") {
+        return null;
+    }
+
+    const functionName = (callee.name || "").toLowerCase();
+    const cachedSuffix = sizeFunctionSuffixes.get(functionName);
+    if (!cachedSuffix) {
+        return null;
+    }
+
+    const args = Array.isArray(callExpression.arguments) ? callExpression.arguments : [];
+    if (args.length !== 1) {
+        return null;
+    }
+
+    const arrayIdentifier = args[0];
+    const arrayIdentifierName = getIdentifierText(arrayIdentifier);
+    if (!arrayIdentifier || !arrayIdentifierName) {
+        return null;
+    }
+
+    const iterator = test.left;
+    if (!iterator || iterator.type !== "Identifier" || !iterator.name) {
+        return null;
+    }
+
+    const update = node.update;
+    if (!update) {
+        return null;
+    }
+
+    if (update.type === "IncDecStatement") {
+        const argument = update.argument;
+        if (!argument || argument.type !== "Identifier" || argument.name !== iterator.name) {
+            return null;
+        }
+    } else if (update.type === "AssignmentExpression") {
+        const left = update.left;
+        if (!left || left.type !== "Identifier" || left.name !== iterator.name) {
+            return null;
+        }
+
+        const allowedOperators = new Set(["+=", "-="]);
+        if (!allowedOperators.has(update.operator)) {
+            return null;
+        }
+    } else {
+        return null;
+    }
+
+    return {
+        iteratorName: iterator.name,
+        sizeIdentifierName: arrayIdentifierName,
+        cachedLengthSuffix: cachedSuffix
+    };
+}
+
+function buildCachedSizeVariableName(baseName, suffix) {
+    const normalizedSuffix = suffix || "len";
+
+    if (!baseName) {
+        return `cached_${normalizedSuffix}`;
+    }
+
+    if (baseName.endsWith(`_${normalizedSuffix}`)) {
+        return baseName;
+    }
+
+    return `${baseName}_${normalizedSuffix}`;
+}
+
+function getIdentifierText(node) {
+    if (!node) {
+        return null;
+    }
+
+    if (node.type === "Identifier") {
+        return node.name || null;
+    }
+
+    if (node.type === "MemberIndexExpression") {
+        const object = node.object;
+        if (!object || object.type !== "Identifier") {
+            return null;
+        }
+
+        if (!Array.isArray(node.property) || node.property.length !== 1) {
+            return null;
+        }
+
+        const indexNode = node.property[0];
+        const indexText = getIdentifierText(indexNode);
+        if (indexText == null) {
+            return null;
+        }
+
+        return `${object.name}_${indexText}`;
+    }
+
+    if (node.type === "MemberDotExpression") {
+        const object = node.object;
+        const property = node.property;
+
+        if (!object || object.type !== "Identifier" || !property || property.type !== "Identifier") {
+            return null;
+        }
+
+        return `${object.name}_${property.name}`;
+    }
+
+    if (node.type === "Literal" && typeof node.value === "string") {
+        return node.value;
+    }
+
+    return null;
+}
+
+export {
+    SIZE_RETRIEVAL_FUNCTION_SUFFIXES,
+    buildCachedSizeVariableName,
+    getArrayLengthHoistInfo
+};
